test(fab): migrate basic
